requirements.txt
annotated-types==0.5.0
anyio==3.7.1
black==23.7.0
click==8.1.7
fastapi==0.101.1
h11==0.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pydantic==2.2.1
pydantic_core==2.6.1
pyhumps==3.8.0
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
pandas
scikit-learn==1.2.2
